requirements.txt
flask_ml>=0.2.0  
onnxruntime==1.19.0
numpy>=2.2.0
torch>=2.5.0
torchvision>=0.20.0
Pillow>=11.0.0